Extract string-builder to be a shared dependency
