+++ src/CMakeLists.txt
@@ -1,3 +1,15 @@
+if (MSVC)
+    add_compile_options(/W4 /WX)
+else()
+    add_compile_options(-Wall -Wextra -pedantic -Werror)
+endif()
+
+if (ON_WINDOWS)
+    if (CMAKE_C_COMPILER_ID MATCHES GNU OR CMAKE_CXX_COMPILER_ID MATCHES GNU)
+        add_compile_options(-Wno-pedantic-ms-format)
+    endif()
+endif()
+
 add_subdirectory(common)
 add_subdirectory(seabolt)
 add_subdirectory(seabolt-cli)
