+++ src/seabolt/src/utils.h
@@ -23,26 +23,6 @@
 #include "config.h"
 #include <time.h>
 
-struct StringBuilder {
-    char* buffer;
-    size_t buffer_size;
-    size_t buffer_pos;
-};
-
-struct StringBuilder* StringBuilder_create();
-
-void StringBuilder_destroy(struct StringBuilder* builder);
-
-void StringBuilder_append(struct StringBuilder* builder, const char* string);
-
-void StringBuilder_append_n(struct StringBuilder* builder, const char* string, const size_t len);
-
-void StringBuilder_append_f(struct StringBuilder* builder, const char* format, ...);
-
-char* StringBuilder_get_string(struct StringBuilder* builder);
-
-size_t StringBuilder_get_length(struct StringBuilder* builder);
-
 void BoltUtil_diff_time(struct timespec* t, struct timespec* t0, struct timespec* t1);
 
 #endif //SEABOLT_ALL_UTILS_H
